Improve display of stdout/stderr for command
Update `git-xargs` to write the `stdout` / `stderr` from the command directly to the logs, as plain text, rather than using structured logging / logrus fields. The problem with using fields, as we did before, is the log output gets shoved into JSON, with all sorts of characters escaped (e.g., newlines become `\n`), which makes it hard to read the log output when debugging.